Speed

+++ src/main/java/mech/mania/engine/game/items/TempStatusModifier.java
@@ -2,11 +2,11 @@ package mech.mania.engine.game.items;
 
 public class TempStatusModifier extends StatusModifier {
     private int duration;
-    private double damagePerTurn;               // Concept: poison type of damage that effects player over time
+    private int damagePerTurn;               // Concept: poison type of damage that effects player over time
 
-    public TempStatusModifier(double speedChange, double healthChange, double experienceChange,
-                              double magicDamageChange, double physicalDamageChange, double magicDefenseChange,
-                              double physicalDefenseChange, int duration, double damagePerTurn) {
+    public TempStatusModifier(int speedChange, int healthChange, int experienceChange,
+                              int magicDamageChange, int physicalDamageChange, int magicDefenseChange,
+                              int physicalDefenseChange, int duration, int damagePerTurn) {
         super(speedChange, healthChange, experienceChange, magicDamageChange, physicalDamageChange, magicDefenseChange,
                 physicalDefenseChange);
         this.duration = duration;
